further ui tests and viewmodel refresh
